added timer module

+++ backend/src/recipes/controllers/recipe.controller.js
@@ -14,7 +14,6 @@ export class RecipeController {
         // bind context
         this.validatePayload = this.validatePayload.bind(this);
         this.getAllRecipes = this.getAllRecipes.bind(this);
-        this.getPrivateRecipes = this.getPrivateRecipes.bind(this);
         this.getRecipesByUserId = this.getRecipesByUserId.bind(this);
         this.getRecipeById = this.getRecipeById.bind(this);
         this.addNewRecipe = this.addNewRecipe.bind(this);
@@ -26,14 +25,17 @@ export class RecipeController {
      * PAJ - Check Cookie header present
      * @param req
      * @param res
-     * @returns user / 401
+     * @returns boolean
      */
-    checkAuthentication (req, res) {
-        if (!this.authService.authenticateUser(req)) {
-            return res.sendStatus(401);
-        } else {
-            return req.user;
-        }
+    validatePayload (req) {
+        return !req.body.title || req.body.title === '' ||
+            !req.body.hasOwnProperty('isPrivate') ||
+            !req.body.description || req.body.description === '' ||
+            !req.body.content || req.body.content === '' ||
+            !req.body.groups || req.body.groups === {} ||
+            !req.body.tags || Array.isArray(req.body.tags) ||
+            !req.body.items || Array.isArray(req.body.items) ||
+            !req.body.timers || Array.isArray(req.body.timers);
     }
 
     /**
@@ -44,8 +46,16 @@ export class RecipeController {
      */
     async getAllRecipes (req, res) {
         try {
-            const result = await this.recipeService.getAllRecipes(false);
-            return res.json(result);
+            const user = this.authService.fetchUserDetails(req);
+            if (user) {
+                console.log(`User logged in, Fetching all recipes`.info);
+                const result = await this.recipeService.getAllRecipes();
+                return res.json(result);
+            } else {
+                console.log(`Fetching all public recipes`.info);
+                const result = await this.recipeService.getPublicRecipes();
+                return res.json(result);
+            }
         } catch (err) {
             console.log(`${this.logger} error fetch all recipes: ${JSON.stringify(err)}`.error);
             return res.sendStatus(500);
@@ -59,7 +69,8 @@ export class RecipeController {
      * @returns {Promise<any>}
      */
     async getRecipesByUserId (req, res) {
-        const user = this.checkAuthentication(req, res);
+        const user = this.authService.fetchUserDetails(req);
+        if (!Boolean(user)) return res.sendStatus(401);
         try {
             const result = await this.recipeService.getAllRecipesByUserId(user.id);
             return res.json(result);
@@ -92,11 +103,12 @@ export class RecipeController {
      * @returns {Promise<any>}
      */
     async addNewRecipe (req, res) {
-        const user = this.checkAuthentication(req, res);
+        const user = this.authService.fetchUserDetails(req);
+        if (!Boolean(user)) return res.sendStatus(401);
         try {
             const result = await this.recipeService.addNewRecipe(req.body, user.id);
             console.log(`${this.logger} - New Record added`, result);
-            return res.sendStatus(201);
+            return res.status(201).send(result._id);
         } catch (err) {
             if (err.code === 11000) {
                 console.log(`${this.logger} Duplicate Record: ${JSON.stringify(err)}`.error);
@@ -117,7 +129,8 @@ export class RecipeController {
      * @returns {Promise<*>}
      */
     async updateRecipeById (req, res) {
-        const user = this.checkAuthentication(req, res);
+        const user = this.authService.fetchUserDetails(req);
+        if (!Boolean(user)) return res.sendStatus(401);
         try {
             const result = await this.recipeService.updateRecipeById(req.params.recipeId, req.body, user.id);
             console.log(`${this.logger} - Record updated: `, result);
@@ -135,7 +148,8 @@ export class RecipeController {
      * @returns {Promise<*>}
      */
     async deleteRecipeById (req, res) {
-        this.checkAuthentication(req, res);
+        const user = this.authService.fetchUserDetails(req);
+        if (!Boolean(user)) return res.sendStatus(401);
         try {
             const result = await this.recipeService.deleteRecipeById(req.params.recipeId);
             console.log(`${this.logger} - Record deleted: `, result);
